trigger once a week
these tests do not need to be run every 6h. The repo itself as a trigger
is only valid when we add new versions of TAS via the values.yml. In between
TAS / Opsman will see minor bumps that (maybe?) could introduce breaking
changes. Triggering once a week on a timer should be enough.