move ci to runway (#86)
* move ci to runway

- organize secrets
- add merge pr job
- add pipeline linter

* trigger once a week

these tests do not need to be run every 6h. The repo itself as a trigger
is only valid when we add new versions of TAS via the values.yml. In between
TAS / Opsman will see minor bumps that (maybe?) could introduce breaking
changes. Triggering once a week on a timer should be enough.

* remediate flakyness by resizing failing vms

the default vm_size on shepherd envs takes too long to execute some steps of
a bbr backup. The bbr backup fails while waiting for the post unlock scripts.

The post unlock scripts unfortunately have a hardcoded timeout of 60 seconds

* add a automated PR to trigger ci once-a-week

this repo  barely gets any commits or PRs. We still want to ensure
that we test against TAS regularly. This commmit achieves that by
creating a PR once a week.

* fix triggers

* add missing trigger