remove usage of harbor, implement adr-006
[#187074138]

remove usage of harbor and ensure each task has an image attribute set
so we can manage ci task images as concourse resources.

switch back ci to run tasks from `bbr-pipeline-tasks` repo with the `image-bbr-pipeline-tasks`.
The whole point of this repo is to have a set of tasks bundled with a compatible image. When the
pipeline for this repo was moved to runway, it was overlooked that executing the pipeline also
should test the underlying docker image ( built specifically for these tasks ).

the dockerfile has been updated so that it's base image can be pulled from an accessible registry
and preloaded ( avoiding implicit docker pulls ) into the oci build task.

fly-by:
- replace docker-image with registry-imgae
- converge on cryogenics-essentials where possible